Réorganisation structurelle du projet -> blocs abstraits
