+++ src/general.py
@@ -78,218 +78,6 @@ class CheckBox(Div):
     def display(self, screen: pygame.Surface):  # -------------------
         screen.blit(self.surface, self.hit_box)
 
-# ============================================================================
-
-
-class Text(Div):
-    """
-    Manage le texte affiché sur une surface pour
-    lui permettre de s'adapter à la taille du conteneur.
-    """
-    def __init__(self, size: tuple, pos: tuple, text: str, font_size: int):
-        super().__init__(size, pos)
-        self.raw_text = text
-        self.font_size = font_size
-
-        # Text for screen
-        self.screen_text = self.wrap(self.raw_text)
-
-    def wrap(self, text: str, origin=(0, 0)) -> list:  # ----------------------------------------------
-        """
-        Gère l'attribution des positions de chaque mots du texte fournit.
-        """
-        # initialisation
-        screen_text = []
-        x, y = origin
-        words = []
-
-        # Words building
-        word = ""
-        for char in text:
-            if char == " ":
-                if word:
-                    words.append(word)
-                    word = ""
-                words.append(char)
-            else:
-                word += char
-        if word:
-            words.append(word)
-
-        # Récuperation de la taille de chaque mot
-        for word in words:
-            screen_word = get_screen_text_for(word, self.font_size)
-            word_rect = screen_word.get_rect()
-
-            # Si c'est le premier mot sur la ligne
-            if x == 0:
-                word_rect.topleft = (x, y)
-                x += word_rect.width
-
-            # Si le mot rentre entièrement sur la ligne
-            elif x + word_rect.width <= self.hit_box.width:
-                word_rect.topleft = (x, y)
-                x += word_rect.width
-
-            # Le mot dépasse de la ligne
-            else:
-                x = 0
-                y += word_rect.height
-                word_rect.topleft = (x, y)
-                x += word_rect.width
-
-            screen_text.append((screen_word, word_rect))
-
-        return screen_text
-
-    def display(self, screen: pygame.Surface):  # -------------------
-        self.surface.fill((0, 0, 0))
-        for elt in self.screen_text:
-            self.surface.blit(elt[0], elt[1])
-        screen.blit(self.surface, self.hit_box)
-
-# ============================================================================
-
-
-class InputVisualizer(Text):
-    """ Permet la visualisation correct d'un champ de texte (input) """
-    def __init__(self, size: tuple, pos: tuple, lines_count: int, left="", right=""):
-        super().__init__(size, pos, left + right, size[1] // lines_count)
-        self.left = left
-        self.right = right
-
-        # Cursor
-        self.cursor = pygame.Surface((3, self.font_size))
-        self.cursor.fill((255, 255, 255))
-        self.cursor_rect = self.cursor.get_rect()
-
-        # Cursor visibility
-        self.time = pygame.time.Clock()
-        self.interval = 500
-        self.reste = 0
-        self.visible = True
-
-    def change_text(self, left="", right=""):  # --------------------
-
-        # réinitialisation clignotement curseur
-        if left != self.left:
-            self.reste = 0
-            self.visible = True
-
-        # Modifications des strings
-        self.left = left
-        self.right = right
-        self.raw_text = self.left + self.right
-        self.wrap_input()
-
-    def wrap_input(self):  # ----------------------------------------
-
-        # Réinitialisation de la pos du curseur
-        self.cursor_rect.topleft = (0, 0)
-
-        # wrap partie gauche curseur
-        self.screen_text = self.wrap(
-            self.left,
-            origin=self.cursor_rect.topleft
-        )
-
-        # Nouvelles coordonnées curseur
-        if self.screen_text:
-            self.cursor_rect.topleft = self.screen_text[-1][1].topright
-
-        # wrap partie droite curseur
-        self.screen_text += self.wrap(
-            self.right,
-            origin=self.cursor_rect.topleft
-        )
-
-        # Clignotement de curseur
-        self.time.tick()
-        self.reste += self.time.get_time()
-        if self.reste > self.interval:
-            self.reste %= self.interval
-            self.visible = not self.visible
-
-        # ajout du curseur dans les éléments du texte
-        if self.visible:
-            self.screen_text.append((self.cursor, self.cursor_rect))
-
-# ============================================================================
-
-
-class Scale(Div):
-    """ Permet la séléction d'une valeur dans [1;max_value] """
-    def __init__(self, size: tuple, pos: tuple, max_value: int, value=1):
-        super().__init__(size, pos)
-        self.max_value = max_value
-        self.value = value
-        self.gap = size[0] // 50
-
-        self.elements = []
-        self.set_elements()
-
-    def set_elements(self):  # --------------------------------------
-        self.elements = []
-
-        # Largeur des éléments
-        grad_width = ((self.hit_box.width - self.gap * (self.max_value - 1))
-                      // self.max_value)
-
-        # Création des graduation dynamiquement
-        x = 0
-        for value in range(1, self.max_value + 1):
-
-            # Rendu visuel en fonction de la valeur actuelle
-            boolean = False
-            if value <= self.value:
-                boolean = True
-
-            # Ajout des éléments à la liste
-            self.elements.append(
-                Graduation(
-                    (grad_width, self.hit_box.height),
-                    (x, 0),
-                    boolean,
-                    value
-                )
-            )
-
-            # Pos
-            x += grad_width + self.gap
-
-    def update(self, rel_pos: tuple):  # ----------------------------
-        for elt in self.elements:
-            if elt.is_under(rel_pos):
-                self.value = elt.value
-        self.set_elements()
-
-    def display(self, screen: pygame.Surface):  # -------------------
-        for elt in self.elements:
-            elt.display(self.surface)
-        screen.blit(self.surface, self.hit_box)
-
-# ============================================================================
-
-
-class Graduation(Div):
-    """ Élément qui constitue Scale """
-    def __init__(self, size: tuple, pos: tuple, active: bool, value: int):
-        super().__init__(size, pos)
-        self.active = active
-        self.value = value
-        self.set_color()
-
-    def set_color(self):  # -----------------------------------------
-        if self.active:
-            self.surface.fill((100, 0, 200))
-        else:
-            self.surface.fill((100, 100, 100))
-
-    def display(self, screen: pygame.Surface):  # -------------------
-        screen.blit(self.surface, self.hit_box)
-
-# ============================================================================
-
 
 def get_screen_text_for(text: str, size: int):
     """ Renvoie un texte sous un format exploitable
